+++ src/eas/easstation.hpp
@@ -22,18 +22,18 @@ public:
                     uint16_t dataPort);
     void deleteHost(uint32_t numberHost);
 
-    bool writeBanOfTimestamps(int32_t idHost, bool isBan);
+    bool writeBanOfTimestamps(int idHost, bool isBan);
     bool writeRegisterOfMasks(int idHost);
     bool writeHardReset(int idMaster);
     bool writePermissionOfMasters(int idMaster, bool isOpen);
     bool writeControlDma(int idHost, bool isOn);
     bool writeStartTaskMaster(int idMaster);
-    bool writeRegisterOfReadData(uint32_t idHost,
+    bool writeRegisterOfReadData(int idHost,
                                  bool isOnFirst,
                                  bool isOnSecond,
                                  bool isOnThird,
                                  bool isOnFourth);
-    bool writePermissionOfData(uint32_t idHost,
+    bool writePermissionOfData(int idHost,
                                bool isOnFirst,
                                bool isOnSecond,
                                bool isOnThird,
@@ -41,10 +41,10 @@ public:
     bool writeClearDataBuffer(int idHost);
     bool writeResetRegister(int idMaster);
     bool writePermissionOfTimer(int idMaster);
-    bool writeSetCoincidence(uint32_t idMaster, uint32_t coincidence);
-    bool writeResetCounter(uint32_t idMaster);
-    bool writeSetTimeGate(uint32_t idMaster, int ns);
-    bool writeActiveChannels(uint32_t idMaster,
+    bool writeSetCoincidence(int idMaster, int coincidence);
+    bool writeResetCounter(int idMaster);
+    bool writeSetTimeGate(int idMaster, int ns);
+    bool writeActiveChannels(int idMaster,
                              bool isOnFirst,
                              bool isOnSecond,
                              bool isOnThird,
@@ -53,15 +53,15 @@ public:
                              bool isOnSixth,
                              bool isOnSeventh,
                              bool isOnEighth);
-    bool writeControlTrigger(uint32_t idMaster, int mode);
-    bool writeProgramTrigger(uint32_t idMaster);
-    bool writeResetVme(uint32_t idMaster);
-    bool writeTime(uint32_t idMaster, int hours, int min, int sec, int ms);
-    bool writeControlRegister(uint32_t idPaa);
-    bool writeThreshold(uint32_t idPaa, uint16_t threshold);
-    bool writeDelay(uint32_t idPaa, uint32_t channel, uint16_t delay);
-    bool writeResetChannel(uint32_t idPaa, uint32_t channel);
-    bool writeSmoothing(uint32_t idPaa, uint16_t smoothing);
+    bool writeControlTrigger(int idMaster, int mode);
+    bool writeProgramTrigger(int idMaster);
+    bool writeResetVme(int idMaster);
+    bool writeTime(int idMaster, int hours, int min, int sec, int ms);
+    bool writeControlRegister(int idPaa);
+    bool writeThreshold(int idPaa, int threshold);
+    bool writeDelay(int idPaa, int channel, int delay);
+    bool writeResetChannel(int idPaa, int channel);
+    bool writeSmoothing(int idPaa, int smoothing);
 
 private:
     IoServicePtr service_;
